requirements.txt
Django==5.0
django-htmx==1.17.2
djangorestframework==3.14.0